change default config to be the main config
